Use a better method for determining when a context start or end is repeated.
This uses the stack position instead of the last term.  Instead of
returning at the beginning of the context tracker, the position of the
context in the stack is checked agains the saved context position.  If
those are the same, then the stack is not modified.

This fixes some cases where the ambiguity markers were causing contexts
to not be started because the second call was returning the new context.

The HeredocInitializer is now the HeredocString and can include the
heredoc body.  This allows for the end matter to be after the heredoc
body instead of at the end of the line.