Correct the array and hash variable tokens.
I don't know where the tree sitter "@\\^" $[A-Z] and "%\\^" $[A-Z] come
from, but those aren't valid.

Also, array and hash variable names have the same restrictions as scalar
variables in that they can not start with a digit.  In addition, the
identifier portion (after the @ or %) may be wrapped in braces.