Fix positioning of commands when a new line starts the content.
For example in

```
[@
	command here
@]
```

The new line in this case is ignored by pgml.  Thus the text starts 1
character after the `[@` instead of immediately thereafter.  So the
text's starting position needs to be saved by the pgml parser so that
the pgml formatting done by the lezer parser can set that position
correctly.