Fix PGML and PG text blocks with no content.
Also enable the eslint rule that checks type imports.

Some basic beginning work on better indentation inside PGML blocks.

Disable base autocompletion on the first line of a PGML or PG text
block.  This seems to be a bug in codemirror/lezer that the base
autocompletion occurs here.  It also means that the PGML or PG text
autocompletion doesn't occur here, and it should.