Don't early terminate a 'balance' block.
This can cause tokens inside the block to be incorrectly terminated
early.  For example, in `[@ '[@ "inner code" @]' @]**` the single quote
starts a 'balance' type block, and the `@]` inside the string tries to
terminate the outer `[@` block.

This early termination code may need further consideration.  It is not
part of PGML and is there to ensure that blocks that are improperly
terminated show up in the tree.  The point is to catch starting to type
a new block inside of another.  For example, say you have already typed
``[` `]``, and then you start to type `[$` inside that.  This code
terminates the outer block so that syntax highlighting continues to
work after the outer block.